requirements.txt
# LLM & Agent Framework
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-ollama

# Vector DB
chromadb>=0.5.0

# Graph DB
neo4j>=5.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
